refactor(store): split store modules
