+++ crates/aube-store/src/lib.rs
@@ -3,122 +3,51 @@ extern crate log;
 
 pub mod dirs;
 
+mod cas;
 mod git;
+mod index;
+mod integrity;
+mod tarball;
 
 pub use git::{
     codeload_cache_lookup, extract_codeload_tarball, git_host_in_list, git_resolve_ref,
     git_shallow_clone, git_url_host,
 };
+
+#[cfg(test)]
+pub(crate) use cas::blake3_hex;
+pub(crate) use cas::cas_file_matches_len;
+use cas::copy_dir_recursive;
 #[cfg(test)]
 use git::{
     codeload_cache_paths, extract_codeload_tarball_at, git_commit_matches, validate_git_positional,
 };
+pub use index::{PackageIndex, StoredFile};
+pub use integrity::{
+    SHA512_INTEGRITY_PREFIX, integrity_to_hex, validate_and_encode_name, validate_pkg_content,
+    validate_version, verify_integrity, verify_precomputed_sha512,
+};
+#[cfg(test)]
+pub(crate) use tarball::normalize_tar_entry_path;
+pub(crate) use tarball::{
+    CappedReader, MAX_TARBALL_DECOMPRESSED_BYTES, MAX_TARBALL_ENTRIES, MAX_TARBALL_ENTRY_BYTES,
+};
 
-use serde::{Deserialize, Serialize};
+#[cfg(test)]
 use sha1::Sha1;
+#[cfg(test)]
 use sha2::{Digest, Sha256, Sha384, Sha512};
-use std::cell::RefCell;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
-use std::sync::atomic::{AtomicBool, Ordering};
-
-pub const SHA512_INTEGRITY_PREFIX: &str = "sha512-";
-
-/// Subresource Integrity (SRI) algorithm prefixes aube accepts in
-/// `dist.integrity`. sha512 is what modern registries emit; sha1 is
-/// kept for legacy packages (e.g. `co@4.6.0`) that were published
-/// before npm's 2017 SRI rollout and never had their metadata rewritten.
-const SRI_PREFIXES: &[(&str, IntegrityAlgo)] = &[
-    ("sha512-", IntegrityAlgo::Sha512),
-    ("sha384-", IntegrityAlgo::Sha384),
-    ("sha256-", IntegrityAlgo::Sha256),
-    ("sha1-", IntegrityAlgo::Sha1),
-];
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-enum IntegrityAlgo {
-    Sha1,
-    Sha256,
-    Sha384,
-    Sha512,
-}
-
-impl IntegrityAlgo {
-    fn prefix(self) -> &'static str {
-        match self {
-            Self::Sha1 => "sha1-",
-            Self::Sha256 => "sha256-",
-            Self::Sha384 => "sha384-",
-            Self::Sha512 => "sha512-",
-        }
-    }
-}
-
-fn parse_sri(expected: &str) -> Option<(IntegrityAlgo, &str)> {
-    SRI_PREFIXES
-        .iter()
-        .find_map(|(prefix, algo)| expected.strip_prefix(prefix).map(|rest| (*algo, rest)))
-}
+use std::sync::atomic::AtomicBool;
+#[cfg(target_os = "macos")]
+use std::sync::atomic::Ordering;
 
 pub const CACHE_DIR_NAME: &str = "aube-cache";
 pub const INDEX_SUBDIR: &str = "index";
 pub const VIRTUAL_STORE_SUBDIR: &str = "virtual-store";
 pub const PACKUMENT_CACHE_SUBDIR: &str = "packuments-v1";
 pub const PACKUMENT_FULL_CACHE_SUBDIR: &str = "packuments-full-v1";
-
-thread_local! {
-    static B3_HASHER: RefCell<blake3::Hasher> = RefCell::new(blake3::Hasher::new());
-    static SHA512_HASHER: RefCell<Sha512> = RefCell::new(Sha512::new());
-}
-
-/// Per-shard mutex array used by the macOS CAS fast path to serialize
-/// concurrent writers within a single process. Indexed by the first
-/// byte of the file's BLAKE3 hash (matching the on-disk 2-char shard
-/// layout), so two threads writing the same hash always collide; threads
-/// writing different hashes typically don't. The array is process-global
-/// rather than per-`Store` because there is at most one active store
-/// per install, and a static avoids carrying 256 mutexes in every cheap
-/// `Store::clone()` along the fetch pipeline.
-///
-/// macOS-gated rather than `not(linux)` because the fast-path block
-/// itself uses `OpenOptionsExt::mode`, which only exists on Unix —
-/// Windows would fail to compile under `not(linux)`. Linux already has
-/// `O_TMPFILE + linkat` (atomic-by-construction, faster than either
-/// alternative); Windows keeps the tempfile + persist_noclobber path.
-#[cfg(target_os = "macos")]
-static FAST_PATH_SHARD_LOCKS: [std::sync::Mutex<()>; 256] =
-    [const { std::sync::Mutex::new(()) }; 256];
-
-/// Recursively copy `src` into `dst`. Used only by the one-shot
-/// legacy-index migration fallback when `rename` fails (typically
-/// cross-filesystem). Not a hot path; correctness > speed.
-fn copy_dir_recursive(src: &Path, dst: &Path) -> std::io::Result<()> {
-    std::fs::create_dir_all(dst)?;
-    for entry in std::fs::read_dir(src)? {
-        let entry = entry?;
-        let file_type = entry.file_type()?;
-        let from = entry.path();
-        let to = dst.join(entry.file_name());
-        if file_type.is_dir() {
-            copy_dir_recursive(&from, &to)?;
-        } else if file_type.is_file() {
-            std::fs::copy(&from, &to)?;
-        }
-        // Symlinks and other types are skipped — the index cache only
-        // ever contains regular JSON files (optionally under a single
-        // level of integrity-shard subdirs).
-    }
-    Ok(())
-}
-
-fn blake3_hex(content: &[u8]) -> String {
-    B3_HASHER.with(|cell| {
-        let mut h = cell.borrow_mut();
-        h.reset();
-        h.update(content);
-        h.finalize().to_hex().to_string()
-    })
-}
 
 /// The global content-addressable store, owned by aube.
 ///
@@ -147,68 +76,6 @@ pub struct Store {
     /// file lock taken at install start. Linux is unaffected because the
     /// O_TMPFILE+linkat path is already atomic-by-construction.
     fast_path: Arc<AtomicBool>,
-}
-
-/// Metadata about a file stored in the CAS.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct StoredFile {
-    /// The hex hash of the file content.
-    pub hex_hash: String,
-    /// The path within the store.
-    pub store_path: PathBuf,
-    /// Whether the file is executable.
-    pub executable: bool,
-    /// File size in bytes when the entry was imported.
-    #[serde(default)]
-    pub size: Option<u64>,
-}
-
-/// Index of all files in a package, keyed by relative path within the package.
-///
-/// Backed by `FxMap` (foldhash) rather than `BTreeMap`: the linker
-/// iterates this map per package and only two non-hot call sites do
-/// keyed lookups (`ignored_builds` checks for `"package.json"` and
-/// `"binding.gyp"`). Hash-based lookup is O(1) for those, and the
-/// flat-bucket layout deserializes/clones with one allocation
-/// instead of one per entry. Iteration order is no longer
-/// lexicographic — cache JSON files now ship in hash order, which
-/// doesn't affect any caller (caches are keyed by tarball path, not
-/// file content).
-pub type PackageIndex = aube_util::collections::FxMap<String, StoredFile>;
-
-fn index_files_match_metadata(index: &PackageIndex, verify_all: bool) -> bool {
-    let mut files = index.values();
-    if verify_all {
-        return files.all(stored_file_matches_metadata);
-    }
-    // Hot install path: one metadata check catches the common crash
-    // residue class (zero-byte/missing CAS files) without turning every
-    // warm lockfile install into a full store walk.
-    files.next().is_none_or(stored_file_matches_metadata)
-}
-
-fn stored_file_matches_metadata(file: &StoredFile) -> bool {
-    file.size
-        .map(|size| cas_file_matches_len(&file.store_path, size))
-        .unwrap_or_else(|| file.store_path.exists())
-}
-
-fn cas_file_matches_len(path: &Path, expected_len: u64) -> bool {
-    path.metadata()
-        .map(|metadata| metadata.len() == expected_len)
-        .unwrap_or(false)
-}
-
-/// Outcome of `create_cas_file`. `Created` means we wrote the bytes
-/// at the final path; `AlreadyExisted` means another writer (or a
-/// previous import) had already committed bit-identical content. The
-/// distinction lets `import_bytes` skip the post-write length check
-/// on the freshly-created path — the file IS exactly the bytes we
-/// just wrote.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum CasWriteOutcome {
-    Created,
-    AlreadyExisted,
 }
 
 impl Store {
@@ -422,1617 +289,6 @@ impl Store {
         let (shard, rest) = hex_hash.split_at(2);
         self.root.join(shard).join(rest)
     }
-
-    /// Load a cached package index, if it exists.
-    ///
-    /// `integrity`, when `Some`, is the registry-advertised SRI
-    /// digest (`sha512-`, or legacy `sha1-` / `sha256-` / `sha384-`)
-    /// of the tarball these cache files came from —
-    /// part of the cache key so the same `(name, version)` resolved
-    /// from different sources (npm registry vs. github codeload vs. a
-    /// proxy that served different bytes) can't alias on disk and
-    /// return each other's file lists to the linker. `None` falls
-    /// back to an unsuffixed `<name>@<version>.json` key so packages
-    /// fetched through a registry proxy that strips `dist.integrity`
-    /// can still warm-install — an integrity-less setup is already a
-    /// degraded mode the user opted into via `strict-store-integrity=false`.
-    pub fn load_index(
-        &self,
-        name: &str,
-        version: &str,
-        integrity: Option<&str>,
-    ) -> Option<PackageIndex> {
-        self.load_index_inner(name, version, integrity, false)
-    }
-
-    /// Load a package index, optionally verifying that all store files still exist.
-    /// The verified variant is slower (stat per file) but detects a corrupted store.
-    pub fn load_index_verified(
-        &self,
-        name: &str,
-        version: &str,
-        integrity: Option<&str>,
-    ) -> Option<PackageIndex> {
-        self.load_index_inner(name, version, integrity, true)
-    }
-
-    fn load_index_inner(
-        &self,
-        name: &str,
-        version: &str,
-        integrity: Option<&str>,
-        verify_files: bool,
-    ) -> Option<PackageIndex> {
-        let index_path = self.index_path(name, version, integrity)?;
-        let buf = xx::file::read(&index_path).ok()?;
-        let index: PackageIndex = sonic_rs::from_slice(&buf).ok()?;
-        if !index_files_match_metadata(&index, verify_files) {
-            trace!("cache stale: {name}@{version}");
-            let _ = xx::file::remove_file(&index_path);
-            return None;
-        }
-        trace!("cache hit: {name}@{version}");
-        Some(index)
-    }
-
-    /// Delete the cached package index for `(name, version, integrity)` if
-    /// it exists. Used as a recovery hatch when the linker discovers a
-    /// CAS shard referenced by the index has gone missing — the cached
-    /// JSON points at a dead `store_path`, so the next install must
-    /// re-derive the index by re-importing the tarball.
-    ///
-    /// `Ok(true)` when an entry was removed; `Ok(false)` when there
-    /// was nothing to remove (or the coordinate was invalid). Errors
-    /// surface only on real I/O failure, not on the missing-file case.
-    pub fn invalidate_cached_index(
-        &self,
-        name: &str,
-        version: &str,
-        integrity: Option<&str>,
-    ) -> Result<bool, Error> {
-        let Some(index_path) = self.index_path(name, version, integrity) else {
-            return Ok(false);
-        };
-        match std::fs::remove_file(&index_path) {
-            Ok(()) => Ok(true),
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => Ok(false),
-            Err(e) => Err(Error::Io(index_path, e)),
-        }
-    }
-
-    /// Save a package index to the cache.
-    ///
-    /// See [`load_index`](Self::load_index) for the semantics of
-    /// `integrity` and the integrity-less fallback.
-    pub fn save_index(
-        &self,
-        name: &str,
-        version: &str,
-        integrity: Option<&str>,
-        index: &PackageIndex,
-    ) -> Result<(), Error> {
-        let index_path = self.index_path(name, version, integrity).ok_or_else(|| {
-            Error::Tar(format!(
-                "refusing to cache: invalid coordinate {name:?}@{version:?} or integrity {integrity:?}"
-            ))
-        })?;
-        let json =
-            serde_json::to_string(index).map_err(|e| Error::Tar(format!("serialize: {e}")))?;
-        xx::file::write(&index_path, json).map_err(|e| Error::Xx(e.to_string()))?;
-        trace!("cached index: {name}@{version}");
-        Ok(())
-    }
-
-    /// Build the on-disk path for a cached index.
-    ///
-    /// Layout:
-    /// - With integrity: `index/<16 hex>/<name>@<version>.json`. The
-    ///   integrity hex lives in a subdirectory (not as part of the
-    ///   filename) so a version whose semver build metadata happens
-    ///   to be 16 lowercase hex chars (e.g. `1.0.0+a1b2c3d4e5f6a7b8`)
-    ///   can never collide with an integrity-keyed entry for
-    ///   `1.0.0` — they land in distinct directories by construction.
-    /// - Without integrity: `index/<name>@<version>.json` at the
-    ///   index dir root. Used for registry proxies that strip
-    ///   `dist.integrity`; the user has already opted out of
-    ///   cross-source integrity enforcement.
-    ///
-    /// Returns `None` when any component is invalid (including an
-    /// integrity string we can't hex-decode).
-    fn index_path(&self, name: &str, version: &str, integrity: Option<&str>) -> Option<PathBuf> {
-        let safe_name = validate_and_encode_name(name)?;
-        if !validate_version(version) {
-            return None;
-        }
-        let filename = format!("{safe_name}@{version}.json");
-        let dir = self.index_dir();
-        match integrity {
-            Some(i) => {
-                let hex = integrity_to_hex(i)?;
-                // 16 hex chars = 64 bits of tarball SHA-512 prefix.
-                // Two tarballs whose SHA-512 prefixes collide would
-                // both have to be valid registry responses for the
-                // same (name, version) *and* survive `verify_integrity`
-                // on fetch, so birthday-bound collisions aren't a
-                // correctness risk; 16 chars is plenty.
-                let short = &hex[..16.min(hex.len())];
-                Some(dir.join(short).join(filename))
-            }
-            None => Some(dir.join(filename)),
-        }
-    }
-
-    /// Ensure every two-char shard directory under the CAS root exists.
-    /// CAS files live under `<root>/<ab>/<cdef...>` for 256 possible
-    /// prefixes. Running this once before a batch of `import_bytes`
-    /// calls lets the per-file hot path skip the `mkdirp(parent)` stat
-    /// entirely (the parent is guaranteed to exist). On APFS that
-    /// removes ~7.5k redundant `stat` syscalls per cold install — the
-    /// `mkdirp` inside `xx::file::write` was the #1 stat hotspot in a
-    /// dtrace profile.
-    ///
-    /// Cheap to call repeatedly: each `create_dir_all` is a no-op when
-    /// the directory already exists, but callers should still hoist the
-    /// call out of tight loops.
-    pub fn ensure_shards_exist(&self) -> Result<(), Error> {
-        std::fs::create_dir_all(&self.root).map_err(|e| Error::Io(self.root.clone(), e))?;
-        // Windows Defender and Search both touch every file in the
-        // store on default installs. Setting this attribute makes
-        // them skip. Non-NTFS volumes ignore it harmlessly.
-        aube_util::fs::set_not_content_indexed(&self.root);
-        let mut buf = [0u8; 2];
-        for hi in 0u8..16 {
-            for lo in 0u8..16 {
-                buf[0] = hex_digit(hi);
-                buf[1] = hex_digit(lo);
-                // SAFETY: every byte in `buf` comes from `hex_digit`,
-                // which only emits `0-9` / `a-f` — always valid UTF-8.
-                let shard = std::str::from_utf8(&buf).unwrap();
-                let path = self.root.join(shard);
-                std::fs::create_dir_all(&path).map_err(|e| Error::Io(path, e))?;
-            }
-        }
-        Ok(())
-    }
-
-    /// Atomically create `path` without overwriting an existing CAS entry.
-    /// `AlreadyExists` is a no-op here; callers that know the expected content
-    /// length must verify it before trusting a reused path. Non-empty files are
-    /// written through a sibling temp file and persisted with no-clobber
-    /// semantics so an interrupted import cannot leave a torn file at the
-    /// content-addressed path. We intentionally do not fsync every CAS file:
-    /// cold installs import tens of thousands of files, and package-index
-    /// loading rejects missing/truncated entries so they can be fetched again.
-    /// `NotFound` means a concurrent prune or a missed `ensure_shards_exist`
-    /// removed the parent shard; recreate it and retry exactly once before
-    /// surfacing.
-    fn create_cas_file(
-        &self,
-        path: &Path,
-        content: Option<&[u8]>,
-    ) -> Result<CasWriteOutcome, Error> {
-        fn do_create_and_write(
-            this: &Store,
-            path: &Path,
-            content: Option<&[u8]>,
-        ) -> Result<CasWriteOutcome, Error> {
-            if let Some(bytes) = content {
-                // O_TMPFILE creates anon file in parent, linkat
-                // publishes atomically. Skips mkstemp uniqueness probe
-                // and post-write fchmod. Docker overlayfs hits the
-                // EOPNOTSUPP fallback. AUBE_DISABLE_O_TMPFILE for
-                // regression cover.
-                #[cfg(target_os = "linux")]
-                {
-                    static O_TMPFILE_DISABLED: std::sync::OnceLock<bool> =
-                        std::sync::OnceLock::new();
-                    let disabled = *O_TMPFILE_DISABLED
-                        .get_or_init(|| std::env::var_os("AUBE_DISABLE_O_TMPFILE").is_some());
-                    if !disabled {
-                        match try_o_tmpfile_publish(path, bytes) {
-                            Ok(outcome) => return Ok(outcome),
-                            Err(OTmpfileFallback::Unsupported) => {}
-                            Err(OTmpfileFallback::Hard(e)) => return Err(e),
-                        }
-                    }
-                }
-
-                // macOS fast path: direct O_CREAT|O_EXCL at the final
-                // content-addressed path, no tempfile dance. Caller (the
-                // install command) flips `fast_path` on only after
-                // acquiring an exclusive store-level lock against other
-                // aube processes. We additionally serialize writers
-                // *within* this process per shard: two threads importing
-                // the same hash (a CAS-dedupe across packages, 35% of
-                // files on dep-heavy graphs like MUI/CodeMirror) would
-                // otherwise both attempt create_new — the loser sees an
-                // EEXIST against the winner's still-empty fd and the
-                // caller's size-mismatch recovery in `import_bytes` would
-                // unlink the file out from under the still-writing
-                // winner. The shard mutex sequences the open+write so the
-                // loser only observes the file at its final size.
-                //
-                // Crashed-predecessor recovery (the unlink+rewrite path
-                // that the slow path defers to `import_bytes`) runs here
-                // while the mutex is still held, so the caller's recovery
-                // can safely no-op for fast-path writes.
-                //
-                // On APFS the fast path is ~2.25x faster than
-                // tempfile+chmod+persist (~64µs/file vs ~145µs/file in
-                // isolation). macOS-gated rather than `not(linux)`
-                // because `OpenOptionsExt::mode` is unix-only — Windows
-                // keeps the tempfile path.
-                #[cfg(target_os = "macos")]
-                if this.fast_path.load(Ordering::Acquire) {
-                    use std::io::Write;
-                    use std::os::unix::fs::OpenOptionsExt;
-
-                    let shard_idx = path
-                        .parent()
-                        .and_then(|p| p.file_name())
-                        .and_then(|s| s.to_str())
-                        .and_then(|s| u8::from_str_radix(s, 16).ok())
-                        .map(|b| b as usize);
-                    // Every path produced by `file_path_from_hex` lives
-                    // under a 2-char hex shard, so this is the contract
-                    // every fast-path caller satisfies today. The assert
-                    // pins the invariant; if a future caller hands in a
-                    // non-CAS path, release builds skip the fast path
-                    // (falling through to the safe tempfile branch)
-                    // rather than do an unsynchronized write that could
-                    // race with another thread on the same hash.
-                    debug_assert!(
-                        shard_idx.is_some(),
-                        "fast-path CAS write to path without a valid hex shard parent: {}",
-                        path.display()
-                    );
-                    if let Some(i) = shard_idx {
-                        // Mutex poisoning is impossible here — the guard
-                        // is dropped at end of scope without us panicking
-                        // inside, so we either return cleanly or propagate
-                        // an `Err` while still releasing the lock. If a
-                        // future caller panics inside, `unwrap_or_else`
-                        // recovers the guard anyway.
-                        let _shard_guard = FAST_PATH_SHARD_LOCKS[i]
-                            .lock()
-                            .unwrap_or_else(|p| p.into_inner());
-
-                        // `OpenOptionsExt::mode(0o644)` is masked by the
-                        // process umask, so a non-default umask (e.g.
-                        // 0o077) would give CAS files 0o600. The
-                        // tempfile path uses `fchmod`, which ignores
-                        // umask. Match it with an explicit
-                        // `set_permissions` so the same store can't end
-                        // up with mixed-mode files depending on which
-                        // path wrote each entry.
-                        use std::os::unix::fs::PermissionsExt;
-                        let force_mode = std::fs::Permissions::from_mode(0o644);
-                        let open_result = std::fs::OpenOptions::new()
-                            .mode(0o644)
-                            .create_new(true)
-                            .write(true)
-                            .open(path);
-                        match open_result {
-                            Ok(mut f) => {
-                                f.set_permissions(force_mode.clone())
-                                    .map_err(|e| Error::Io(path.to_path_buf(), e))?;
-                                f.write_all(bytes)
-                                    .map_err(|e| Error::Io(path.to_path_buf(), e))?;
-                                return Ok(CasWriteOutcome::Created);
-                            }
-                            Err(e) if e.kind() == std::io::ErrorKind::AlreadyExists => {
-                                // Holding the shard lock, so any in-process
-                                // writer for this hash already finished. If
-                                // the file size matches, it's a genuine
-                                // dedupe. If not, it's a crashed-predecessor
-                                // remnant — unlink and rewrite inline.
-                                if cas_file_matches_len(path, bytes.len() as u64) {
-                                    return Ok(CasWriteOutcome::AlreadyExisted);
-                                }
-                                let _ = xx::file::remove_file(path);
-                                match std::fs::OpenOptions::new()
-                                    .mode(0o644)
-                                    .create_new(true)
-                                    .write(true)
-                                    .open(path)
-                                {
-                                    Ok(mut f) => {
-                                        f.set_permissions(force_mode)
-                                            .map_err(|e| Error::Io(path.to_path_buf(), e))?;
-                                        f.write_all(bytes)
-                                            .map_err(|e| Error::Io(path.to_path_buf(), e))?;
-                                        return Ok(CasWriteOutcome::Created);
-                                    }
-                                    Err(e) => {
-                                        return Err(Error::Io(path.to_path_buf(), e));
-                                    }
-                                }
-                            }
-                            Err(e) if e.kind() == std::io::ErrorKind::NotFound => {
-                                // Shard dir missing — fall through to the slow
-                                // path; the outer wrapper will create_dir_all
-                                // and retry once.
-                            }
-                            Err(e) => return Err(Error::Io(path.to_path_buf(), e)),
-                        }
-                    }
-                }
-
-                // Tempfile + persist_noclobber gives atomic crash
-                // semantics: a partial write on `tmp` is dropped by
-                // tempfile's Drop impl, so the final path either
-                // contains the complete bytes or doesn't exist. A
-                // direct O_CREAT|O_EXCL write to the final path was
-                // tried (faster path, ~3 syscalls per file) but
-                // raced with concurrent installs in CI where two
-                // processes saw the same partial file in different
-                // orders and clobbered each other's recovery. The
-                // fast-path branch above re-enables it under an
-                // exclusive store lock.
-                let _ = this; // suppress unused warning on Linux
-                let parent = path.parent().ok_or_else(|| {
-                    Error::Io(path.to_path_buf(), std::io::ErrorKind::NotFound.into())
-                })?;
-                let mut tmp = tempfile::Builder::new()
-                    .prefix(".aube-cas-")
-                    .tempfile_in(parent)
-                    .map_err(|e| Error::Io(path.to_path_buf(), e))?;
-                use std::io::Write;
-                tmp.write_all(bytes)
-                    .map_err(|e| Error::Io(path.to_path_buf(), e))?;
-                #[cfg(unix)]
-                {
-                    use std::os::unix::fs::PermissionsExt;
-                    tmp.as_file()
-                        .set_permissions(std::fs::Permissions::from_mode(0o644))
-                        .map_err(|e| Error::Io(path.to_path_buf(), e))?;
-                }
-                return match tmp.persist_noclobber(path) {
-                    Ok(_) => Ok(CasWriteOutcome::Created),
-                    Err(e) if e.error.kind() == std::io::ErrorKind::AlreadyExists => {
-                        Ok(CasWriteOutcome::AlreadyExisted)
-                    }
-                    Err(e) => Err(Error::Io(path.to_path_buf(), e.error)),
-                };
-            }
-
-            match std::fs::OpenOptions::new()
-                .write(true)
-                .create_new(true)
-                .open(path)
-            {
-                Ok(_) => Ok(CasWriteOutcome::Created),
-                Err(e) if e.kind() == std::io::ErrorKind::AlreadyExists => {
-                    Ok(CasWriteOutcome::AlreadyExisted)
-                }
-                Err(e) => Err(Error::Io(path.to_path_buf(), e)),
-            }
-        }
-
-        match do_create_and_write(self, path, content) {
-            Ok(outcome) => Ok(outcome),
-            Err(Error::Io(_, ref ioe)) if ioe.kind() == std::io::ErrorKind::NotFound => {
-                // Shard dir missing. `ensure_shards_exist` normally
-                // pre-creates all 256 shards; this only fires when the
-                // caller didn't call it or a concurrent prune wiped
-                // the tree mid-install.
-                if let Some(parent) = path.parent() {
-                    std::fs::create_dir_all(parent)
-                        .map_err(|e| Error::Io(parent.to_path_buf(), e))?;
-                }
-                do_create_and_write(self, path, content)
-            }
-            Err(e) => Err(e),
-        }
-    }
-
-    /// Import a single file's content into the store. Returns the stored file info.
-    ///
-    /// Hot path on cold installs: callers should invoke
-    /// [`Store::ensure_shards_exist`] once before a batch of imports so
-    /// this function can skip the per-file `mkdirp`. When shards don't
-    /// exist yet, the `create_new` open will fail with `NotFound`; we
-    /// fall back to the slow path for correctness.
-    pub fn import_bytes(&self, content: &[u8], executable: bool) -> Result<StoredFile, Error> {
-        let hash_t0 = std::time::Instant::now();
-        let hex_hash = blake3_hex(content);
-        if aube_util::diag::enabled() {
-            aube_util::diag::event_lazy(
-                aube_util::diag::Category::Store,
-                "blake3_hash",
-                hash_t0.elapsed(),
-                || format!(r#"{{"size":{}}}"#, content.len()),
-            );
-        }
-
-        let store_path = self.file_path_from_hex(&hex_hash);
-        let _diag_write =
-            aube_util::diag::Span::new(aube_util::diag::Category::Store, "import_bytes_write")
-                .with_meta_fn(|| format!(r#"{{"size":{}}}"#, content.len()));
-
-        // Fast path: open-with-create-new combines the existence check
-        // and the open into a single syscall. On a cold CAS this does
-        // one open(O_CREAT|O_EXCL|O_WRONLY) per file and replaces the
-        // previous stat+create pair (~15k redundant stats per cold
-        // install). On a warm CAS, concurrent writers are safe: EEXIST
-        // means another writer already materialized this content (same
-        // hash = same bytes), so we skip and share the entry.
-        //
-        // `Created` means we just wrote the bytes — they are exactly
-        // `content.len()` by construction, no need to re-stat. Only
-        // the `AlreadyExisted` branch can produce a torn file (from a
-        // crashed predecessor) so the length check runs there only.
-        let outcome = self.create_cas_file(&store_path, Some(content))?;
-        // Surface CAS dedup hit/miss to diag so cold vs warm vs partial
-        // installs can be classified post-hoc. `cas_hit` fires when an
-        // identical-content file already lived in the store; `cas_miss`
-        // fires when we just wrote new bytes.
-        if aube_util::diag::enabled() {
-            let name = match outcome {
-                CasWriteOutcome::Created => "cas_miss",
-                CasWriteOutcome::AlreadyExisted => "cas_hit",
-            };
-            aube_util::diag::instant_lazy(aube_util::diag::Category::Store, name, || {
-                format!(r#"{{"size":{}}}"#, content.len())
-            });
-        }
-        // The macOS fast path verifies the file size inline under its
-        // shard mutex before returning `AlreadyExisted`, so this
-        // recovery only needs to run when we took the tempfile path.
-        // Skipping it there also prevents a race where the recovery
-        // unlinks a file that another in-process thread is concurrently
-        // re-creating after observing the same crashed-predecessor.
-        //
-        // `cfg!(target_os = "macos")` matches the cfg gate on the only
-        // code path that flips `fast_path` to true (and on the inline
-        // recovery inside `create_cas_file`). Without the cfg!, a future
-        // caller setting the flag on Linux would silently disable this
-        // recovery — the Linux O_TMPFILE branch has no inline
-        // length-check substitute, so torn CAS files would be accepted.
-        let fast_path_handled_recovery =
-            cfg!(target_os = "macos") && self.fast_path.load(Ordering::Acquire);
-        if outcome == CasWriteOutcome::AlreadyExisted && !fast_path_handled_recovery {
-            // A length mismatch from this branch can mean either
-            //   (a) a crashed predecessor left a torn file (the recovery
-            //       case this code was originally written for), or
-            //   (b) on macOS, another *process* is currently writing to
-            //       the same path via the fast path (no atomic publish
-            //       at the final path, so its in-progress fd is visible
-            //       by name to other writers).
-            // Burning the file in case (b) would unlink the active
-            // writer's inode and trigger a cascading recovery race. Wait
-            // briefly (50ms is dozens of typical small-file writes) for
-            // the partial file to settle. If it stays mismatched past
-            // the deadline, treat it as (a) and recover.
-            if !cas_file_matches_len(&store_path, content.len() as u64) {
-                let deadline = std::time::Instant::now() + std::time::Duration::from_millis(50);
-                while !cas_file_matches_len(&store_path, content.len() as u64)
-                    && std::time::Instant::now() < deadline
-                {
-                    std::thread::sleep(std::time::Duration::from_micros(250));
-                }
-            }
-            if !cas_file_matches_len(&store_path, content.len() as u64) {
-                let _ = xx::file::remove_file(&store_path);
-                self.create_cas_file(&store_path, Some(content))?;
-                if !cas_file_matches_len(&store_path, content.len() as u64) {
-                    let actual_len = store_path.metadata().map(|metadata| metadata.len()).ok();
-                    return Err(Error::Io(
-                        store_path.clone(),
-                        std::io::Error::other(format!(
-                            "CAS entry has wrong size after import: expected {} bytes, got {}",
-                            content.len(),
-                            actual_len
-                                .map(|len| format!("{len} bytes"))
-                                .unwrap_or_else(|| "missing file".to_owned())
-                        )),
-                    ));
-                }
-            }
-        }
-
-        if executable {
-            // Behavior note: this branch now runs unconditionally when
-            // `executable=true`, including when the content file
-            // already existed (`AlreadyExists` above). Previously the
-            // marker was only written in the fresh-content branch.
-            // The new shape is strictly more correct — if the same
-            // bytes are imported twice, once with `executable=false`
-            // and once with `true`, the marker should exist after the
-            // second call. Auditing the callers of the `-exec` marker:
-            //   - `aube-store::import_bytes` (this function, the only
-            //     writer).
-            //   - `aube-store` tests (assert the marker exists after
-            //     an `executable=true` import).
-            //   - `aube::commands::store` (`aube store prune`)
-            //     uses the marker to skip bumping the "freed bytes"
-            //     counter when unlinking exec-marker sidecars.
-            // No code path reads the marker to decide executability —
-            // that's carried in `StoredFile.executable`, threaded
-            // through the `PackageIndex` and the linker. So flipping
-            // a marker-absent-to-present for a shared hash is safe.
-            let exec_marker = PathBuf::from(format!("{}-exec", store_path.display()));
-            self.create_cas_file(&exec_marker, None)?;
-        }
-
-        Ok(StoredFile {
-            hex_hash,
-            store_path,
-            executable,
-            size: Some(content.len() as u64),
-        })
-    }
-
-    /// Import every file under a directory into the store, producing a
-    /// `PackageIndex` keyed by paths relative to `dir`. Used by `file:`
-    /// deps pointing at an on-disk package directory. Common noise
-    /// (`.git`, `node_modules`) is skipped so local packages don't drag
-    /// the target's own installed deps into the virtual store.
-    pub fn import_directory(&self, dir: &Path) -> Result<PackageIndex, Error> {
-        let mut index = PackageIndex::default();
-        self.import_directory_recursive(dir, dir, &mut index)?;
-        Ok(index)
-    }
-
-    fn import_directory_recursive(
-        &self,
-        base: &Path,
-        current: &Path,
-        index: &mut PackageIndex,
-    ) -> Result<(), Error> {
-        let entries = std::fs::read_dir(current)
-            .map_err(|e| Error::Tar(format!("read_dir {}: {e}", current.display())))?;
-        for entry in entries {
-            let entry =
-                entry.map_err(|e| Error::Tar(format!("read_dir {}: {e}", current.display())))?;
-            let file_type = entry
-                .file_type()
-                .map_err(|e| Error::Tar(format!("file_type: {e}")))?;
-            let name_os = entry.file_name();
-            let name_str = name_os.to_string_lossy();
-            if matches!(name_str.as_ref(), ".git" | "node_modules") {
-                continue;
-            }
-            let path = entry.path();
-            if file_type.is_dir() {
-                self.import_directory_recursive(base, &path, index)?;
-                continue;
-            }
-            if !file_type.is_file() {
-                continue;
-            }
-            let content = std::fs::read(&path)
-                .map_err(|e| Error::Tar(format!("read {}: {e}", path.display())))?;
-            #[cfg(unix)]
-            let executable = {
-                use std::os::unix::fs::PermissionsExt;
-                let meta = entry
-                    .metadata()
-                    .map_err(|e| Error::Tar(format!("metadata: {e}")))?;
-                meta.permissions().mode() & 0o111 != 0
-            };
-            #[cfg(not(unix))]
-            let executable = false;
-            let stored = self.import_bytes(&content, executable)?;
-            let rel = path
-                .strip_prefix(base)
-                .map_err(|e| Error::Tar(format!("strip_prefix: {e}")))?
-                .to_string_lossy()
-                .replace('\\', "/");
-            index.insert(rel, stored);
-        }
-        Ok(())
-    }
-
-    /// Import a tarball (.tgz) into the store.
-    /// Returns a PackageIndex mapping relative paths to stored files.
-    ///
-    /// Two-phase: serial tar walk that stages
-    /// `(rel_path, content, executable)` triples (the tar reader is
-    /// inherently sequential), then a CAS-write batch. When the
-    /// staged batch crosses [`PARALLEL_IMPORT_THRESHOLD`] entries,
-    /// the writes fan out via `rayon::par_iter` — the per-file CAS
-    /// path is `O_CREAT|O_EXCL` and uses a shared `&Store`, so
-    /// parallel writers are race-safe by construction (`EEXIST` on
-    /// content collision is a success path because BLAKE3 paths are
-    /// content-addressed).
-    ///
-    /// `AUBE_DISABLE_PARALLEL_IMPORT=1` forces the serial path. Use
-    /// it as a regression killswitch if a future rayon scope inversion
-    /// (linker symlink pass running concurrently) shows contention.
-    /// Below the threshold the small-tarball overhead of rayon
-    /// dispatch outweighs the win, so the cutover is conditional.
-    pub fn import_tarball(&self, tarball_bytes: &[u8]) -> Result<PackageIndex, Error> {
-        // &[u8] impls std::io::Read by advancing the slice.
-        self.import_tarball_reader(tarball_bytes)
-    }
-
-    /// Streaming variant. Accepts any compressed-tarball Read source so
-    /// callers can pipe HTTP body chunks straight through without
-    /// buffering the whole archive into memory first. Caps and CAS
-    /// publish semantics match `import_tarball` exactly.
-    pub fn import_tarball_reader<R: std::io::Read>(
-        &self,
-        compressed_reader: R,
-    ) -> Result<PackageIndex, Error> {
-        use std::io::Read;
-
-        let _diag =
-            aube_util::diag::Span::new(aube_util::diag::Category::Store, "import_tarball_reader");
-        let _diag_decode = aube_util::diag::inflight(aube_util::diag::Slot::Decode);
-        let extract_t0 = std::time::Instant::now();
-
-        // Caps defend against gzip bombs and lying tar headers. The
-        // values sit well above any real npm package (largest top
-        // 1000 are in the tens of MiB) but low enough to prevent a
-        // malicious registry or mirror from OOMing the installer
-        // with a small high-compression-ratio payload.
-        //
-        // CappedReader instead of Read::take for the archive-level
-        // cap so exhaustion surfaces as an Err. A clean EOF landing on
-        // a tar block boundary would let a crafted archive silently
-        // truncate into a partial index.
-        let gz = flate2::read::GzDecoder::new(compressed_reader);
-        let capped = CappedReader::new(gz, MAX_TARBALL_DECOMPRESSED_BYTES);
-        let buffered = std::io::BufReader::with_capacity(256 * 1024, capped);
-        let mut archive = tar::Archive::new(buffered);
-        /*
-         * Chunked staged pipeline. Read N entries, flush them to CAS
-         * via rayon parallel writes, repeat. Keeps the existing
-         * rayon global pool warm across chunks and partially
-         * overlaps tar parsing with file writes within a single
-         * tarball. No new threads spawned (per-call thread::scope
-         * was tried and live locked at 80 s, see git history if
-         * curious). Chunk size of 64 is roughly the median npm
-         * package's file count, so most tarballs flush at most
-         * once or twice; fat native bindings (next, sharp, swc)
-         * with 1k+ files chunk through 16+ flushes. The legacy
-         * "stage everything then flush" path remains under
-         * `AUBE_DISABLE_PIPELINED_IMPORT=1` for byte-identity
-         * regression debugging.
-         */
-        const PIPELINE_CHUNK_SIZE: usize = 64;
-        let pipelined_disabled = std::env::var_os("AUBE_DISABLE_PIPELINED_IMPORT").is_some();
-        let parallel_disabled = std::env::var_os("AUBE_DISABLE_PARALLEL_IMPORT").is_some();
-        let mut staged: Vec<(String, Vec<u8>, bool)> = Vec::new();
-        let mut entries_seen: usize = 0;
-        let mut total_uncompressed: u64 = 0;
-        let mut decode_ns: u128 = 0;
-        let mut cas_ns: u128 = 0;
-        let mut index = PackageIndex::default();
-        let mut staged_count: usize = 0;
-
-        let flush_chunk = |chunk: Vec<(String, Vec<u8>, bool)>,
-                           index: &mut PackageIndex,
-                           cas_ns: &mut u128|
-         -> Result<(), Error> {
-            if chunk.is_empty() {
-                return Ok(());
-            }
-            let chunk_t0 = std::time::Instant::now();
-            if parallel_disabled || chunk.len() < PARALLEL_IMPORT_THRESHOLD {
-                for (rel_path, content, executable) in chunk {
-                    let stored = self.import_bytes(&content, executable)?;
-                    index.insert(rel_path, stored);
-                }
-            } else {
-                use rayon::iter::{
-                    IndexedParallelIterator, IntoParallelIterator, ParallelIterator,
-                };
-                // `with_min_len` raises the minimum work unit per
-                // rayon task. samply on a 1230-pkg cold install
-                // pinned `crossbeam_deque::Stealer::steal` at 4.1%
-                // self time; each per-file task is ~50µs of useful
-                // work, below rayon's amortization threshold for
-                // its work-stealing overhead. Grouping 8 files per
-                // task amortizes the dispatch/steal cost without
-                // losing meaningful parallelism — 8 × 50µs = 400µs,
-                // well under a typical OS scheduling slice.
-                const RAYON_TASK_MIN_LEN: usize = 8;
-                let results: Vec<Result<(String, StoredFile), Error>> = chunk
-                    .into_par_iter()
-                    .with_min_len(RAYON_TASK_MIN_LEN)
-                    .map(|(rel_path, content, executable)| {
-                        self.import_bytes(&content, executable)
-                            .map(|stored| (rel_path, stored))
-                    })
-                    .collect();
-                for r in results {
-                    let (rel_path, stored) = r?;
-                    index.insert(rel_path, stored);
-                }
-            }
-            *cas_ns += chunk_t0.elapsed().as_nanos();
-            Ok(())
-        };
-
-        for entry in archive.entries().map_err(|e| Error::Tar(e.to_string()))? {
-            entries_seen += 1;
-            if entries_seen > MAX_TARBALL_ENTRIES {
-                return Err(Error::Tar(format!(
-                    "tarball exceeds entry cap of {MAX_TARBALL_ENTRIES}"
-                )));
-            }
-
-            let mut entry = entry.map_err(|e| Error::Tar(e.to_string()))?;
-
-            // Directories don't carry content, skip them. PAX global
-            // and extension headers (type `g` / `x`) carry metadata
-            // only — GitHub-generated tarballs (e.g. `imap@0.8.19`)
-            // start with one that embeds the source git blob SHA.
-            // npm/pnpm/bun tolerate these; we do too. Every other
-            // non-regular entry type (symlink, hardlink, character
-            // device, block device, fifo) is rejected. Real npm
-            // packages ship files and directories only. Symlink and
-            // hardlink entries are the load-bearing primitive of the
-            // node-tar CVE-2021-37701 class and have no legitimate
-            // use here.
-            let entry_type = entry.header().entry_type();
-            // GNU LongName/LongLink and PAX X-headers carry metadata
-            // for the next real entry. The tar crate folds the long
-            // name into Entry::path() automatically. Just skip the
-            // metadata records themselves.
-            if entry_type.is_dir()
-                || matches!(
-                    entry_type,
-                    tar::EntryType::XGlobalHeader
-                        | tar::EntryType::XHeader
-                        | tar::EntryType::GNULongName
-                        | tar::EntryType::GNULongLink
-                )
-            {
-                continue;
-            }
-            if !matches!(
-                entry_type,
-                tar::EntryType::Regular | tar::EntryType::Continuous
-            ) {
-                return Err(Error::Tar(format!(
-                    "tarball entry type {entry_type:?} is not allowed"
-                )));
-            }
-
-            // Reject oversized entries up front on the declared size
-            // so we never allocate a huge `Vec` just to error after.
-            // `.take()` below is the belt-and-suspenders guard for
-            // the case where the header lies about the stream length.
-            let declared = entry
-                .header()
-                .size()
-                .map_err(|e| Error::Tar(e.to_string()))?;
-            if declared > MAX_TARBALL_ENTRY_BYTES {
-                return Err(Error::Tar(format!(
-                    "tarball entry exceeds per-entry cap: {declared} bytes > {MAX_TARBALL_ENTRY_BYTES}"
-                )));
-            }
-
-            let raw_path = entry
-                .path()
-                .map_err(|e| Error::Tar(e.to_string()))?
-                .to_path_buf();
-            let Some(rel_path) = normalize_tar_entry_path(&raw_path)? else {
-                // Entry was the wrapper directory itself with no
-                // interior path after stripping. Nothing to store.
-                continue;
-            };
-
-            // Clamp upfront alloc so a lying header can't force a 512
-            // MiB reservation before any byte has been read. read_to_end
-            // grows the Vec for the rare entry that really is huge.
-            let mut content = Vec::with_capacity((declared as usize).min(VEC_PREALLOC_CEILING));
-            let read_t0 = std::time::Instant::now();
-            (&mut entry)
-                .take(MAX_TARBALL_ENTRY_BYTES)
-                .read_to_end(&mut content)
-                .map_err(|e| Error::Tar(e.to_string()))?;
-            decode_ns += read_t0.elapsed().as_nanos();
-
-            // Reject header that declared 0 bytes but produced a
-            // non-empty stream. Synthetic-entry injection: header
-            // claims empty file, real bytes go to disk.
-            if declared == 0 && !content.is_empty() {
-                return Err(Error::Tar(format!(
-                    "tarball entry declared 0 bytes but yielded {} bytes",
-                    content.len()
-                )));
-            }
-
-            let mode = entry.header().mode().unwrap_or(0o644);
-            let executable = mode & 0o111 != 0;
-            total_uncompressed = total_uncompressed.saturating_add(content.len() as u64);
-            staged.push((rel_path, content, executable));
-            staged_count += 1;
-
-            if !pipelined_disabled && staged.len() >= PIPELINE_CHUNK_SIZE {
-                let chunk = std::mem::take(&mut staged);
-                flush_chunk(chunk, &mut index, &mut cas_ns)?;
-            }
-        }
-
-        aube_util::diag::event_lazy(
-            aube_util::diag::Category::Store,
-            "tar_extract_complete",
-            extract_t0.elapsed(),
-            || format!(r#"{{"entries":{staged_count},"bytes_uncompressed":{total_uncompressed}}}"#),
-        );
-        if aube_util::diag::enabled() {
-            aube_util::diag::event_lazy(
-                aube_util::diag::Category::Store,
-                "gzip_decompress",
-                std::time::Duration::from_nanos(decode_ns as u64),
-                || format!(r#"{{"bytes_uncompressed":{total_uncompressed}}}"#),
-            );
-        }
-
-        if !staged.is_empty() {
-            let chunk = std::mem::take(&mut staged);
-            flush_chunk(chunk, &mut index, &mut cas_ns)?;
-        }
-        aube_util::diag::event_lazy(
-            aube_util::diag::Category::Store,
-            "cas_import_complete",
-            // Saturating cast: u128 cas_ns won't realistically
-            // exceed u64::MAX (~584 years in nanoseconds), but a
-            // bug or runaway accumulator should clamp to the diag
-            // ceiling rather than silently truncate the high bits
-            // and emit a misleadingly small duration.
-            std::time::Duration::from_nanos(u64::try_from(cas_ns).unwrap_or(u64::MAX)),
-            || {
-                let pipelined = !pipelined_disabled;
-                let parallel = !parallel_disabled && staged_count >= PARALLEL_IMPORT_THRESHOLD;
-                format!(
-                    r#"{{"files":{staged_count},"parallel":{parallel},"pipelined":{pipelined}}}"#
-                )
-            },
-        );
-        Ok(index)
-    }
-}
-
-// Median npm tarball has 7 files. Old 256 threshold almost never
-// tripped. Rayon dispatch is cheap on tiny batches.
-// AUBE_DISABLE_PARALLEL_IMPORT kills the parallel path entirely.
-const PARALLEL_IMPORT_THRESHOLD: usize = 16;
-
-/// Strip the wrapper directory from `raw` and return a safe POSIX-style
-/// index key, or refuse the entry outright.
-///
-/// Rejects every shape that would let a crafted tarball place the
-/// eventual `pkg_dir.join(key)` file outside the package root:
-/// `..` anywhere in the remaining path, absolute paths, Windows
-/// drive prefixes, backslash separators smuggled inside a single
-/// component, and NUL bytes. Non-UTF-8 paths are also rejected because
-/// the stored index is a JSON map keyed by string.
-///
-/// `Ok(None)` means the entry stripped down to the wrapper itself
-/// and should be skipped by the caller.
-///
-/// Matches the class of defences node-tar added for CVE-2021-32804,
-/// CVE-2021-37713, and what pnpm added for CVE-2024-27298.
-fn normalize_tar_entry_path(raw: &Path) -> Result<Option<String>, Error> {
-    use std::path::Component;
-
-    // Peek past any leading `.` segments (`./package/foo.js` appears
-    // in some tar implementations' wrapper representations) so the
-    // first-component reject and the wrapper-strip both work off the
-    // same "first real" position. Running the reject before the
-    // stripping loop would otherwise let `./../file` silently
-    // consume the `..` as the wrapper.
-    let mut components = raw.components().peekable();
-    while matches!(components.peek(), Some(Component::CurDir)) {
-        components.next();
-    }
-
-    // Reject absolute, drive-prefixed, or `..`-rooted paths before
-    // wrapper-strip runs. A naive "skip the first component" would
-    // otherwise strip the `RootDir` marker and accept `/etc` as
-    // `etc`, or consume a leading `..` as the wrapper.
-    match components.peek() {
-        Some(Component::RootDir) => {
-            return Err(Error::Tar(format!(
-                "tarball entry path is absolute: {raw:?}"
-            )));
-        }
-        Some(Component::Prefix(_)) => {
-            return Err(Error::Tar(format!(
-                "tarball entry path has a Windows drive prefix: {raw:?}"
-            )));
-        }
-        Some(Component::ParentDir) => {
-            return Err(Error::Tar(format!(
-                "tarball entry path escapes package root via `..`: {raw:?}"
-            )));
-        }
-        _ => {}
-    }
-
-    // Drop the first real component as the wrapper directory. npm
-    // convention is `package/`, but some packages ship the package
-    // name or another identifier. Whatever it is, drop it.
-    components.next();
-
-    // Pre-size to the raw path length so growing the output never
-    // reallocates: the normalized form drops the wrapper segment and
-    // converts `\` to `/` but never grows beyond the input size.
-    let mut out = String::with_capacity(raw.as_os_str().len());
-    for comp in components {
-        match comp {
-            Component::Normal(os) => {
-                let s = os.to_str().ok_or_else(|| {
-                    Error::Tar(format!(
-                        "tarball entry path contains non-UTF-8 bytes: {raw:?}"
-                    ))
-                })?;
-                if s.is_empty() || s.contains('\0') || s.contains('\\') || s.contains('/') {
-                    return Err(Error::Tar(format!(
-                        "tarball entry path contains a malformed component: {raw:?}"
-                    )));
-                }
-                // Windows-only filename restrictions. Gated to
-                // cfg(windows) so Unix hosts keep tarballs with
-                // valid-on-Linux names like `CON.js` or `foo.`.
-                // Rejecting those cross-platform would regress real
-                // Linux installs for a hazard that only hits
-                // Windows users. Windows users get the checks they
-                // need, portability of a package to Windows is the
-                // publisher's problem to validate.
-                #[cfg(windows)]
-                {
-                    // `:` is an alternate data stream separator on
-                    // NTFS and is rejected by Windows path creation.
-                    if s.contains(':') {
-                        return Err(Error::Tar(format!(
-                            "tarball entry path contains a malformed component: {raw:?}"
-                        )));
-                    }
-                    // NTFS reserved device names. `CON`, `con.txt`,
-                    // `CON.tar.gz` all resolve to the console
-                    // device. Writing one either fails with
-                    // ERROR_INVALID_NAME or gets silently consumed
-                    // by the device driver and hangs the writer.
-                    if is_windows_reserved_name(s) {
-                        return Err(Error::Tar(format!(
-                            "tarball entry path contains a Windows reserved device name: {raw:?}"
-                        )));
-                    }
-                    // NTFS strips trailing `.` and trailing space
-                    // on create so `foo` and `foo.` alias. Reject
-                    // both rather than sort out aliasing at
-                    // materialize time.
-                    if s.ends_with('.') || s.ends_with(' ') {
-                        return Err(Error::Tar(format!(
-                            "tarball entry path has a trailing dot or space which Windows strips: {raw:?}"
-                        )));
-                    }
-                    // Control chars 0x01..0x1F invalid on NTFS.
-                    // Reject the whole tarball instead of hitting
-                    // per-file create errors mid-extract.
-                    if s.bytes().any(|b| b < 0x20) {
-                        return Err(Error::Tar(format!(
-                            "tarball entry path contains control characters: {raw:?}"
-                        )));
-                    }
-                }
-                if !out.is_empty() {
-                    out.push('/');
-                }
-                out.push_str(s);
-            }
-            Component::ParentDir => {
-                return Err(Error::Tar(format!(
-                    "tarball entry path escapes package root via `..`: {raw:?}"
-                )));
-            }
-            Component::RootDir => {
-                return Err(Error::Tar(format!(
-                    "tarball entry path is absolute: {raw:?}"
-                )));
-            }
-            Component::Prefix(_) => {
-                return Err(Error::Tar(format!(
-                    "tarball entry path has a Windows drive prefix: {raw:?}"
-                )));
-            }
-            // `.` components are harmless and appear in some tar
-            // implementations' wrapper representations.
-            Component::CurDir => {}
-        }
-    }
-
-    if out.is_empty() {
-        Ok(None)
-    } else {
-        Ok(Some(out))
-    }
-}
-
-/// Check if a tarball path component matches a Windows reserved
-/// device name. Compare case-insensitively on the stem only.
-/// `CON`, `Con`, `con`, `con.txt`, `CON.tar.gz` all resolve to the
-/// same DOS device on NTFS. Only base name matters, the extension
-/// is irrelevant to the device lookup.
-#[cfg(windows)]
-fn is_windows_reserved_name(name: &str) -> bool {
-    let stem = name.split_once('.').map(|(a, _)| a).unwrap_or(name);
-    let upper = stem.to_ascii_uppercase();
-    matches!(
-        upper.as_str(),
-        "CON"
-            | "PRN"
-            | "AUX"
-            | "NUL"
-            | "COM1"
-            | "COM2"
-            | "COM3"
-            | "COM4"
-            | "COM5"
-            | "COM6"
-            | "COM7"
-            | "COM8"
-            | "COM9"
-            | "LPT1"
-            | "LPT2"
-            | "LPT3"
-            | "LPT4"
-            | "LPT5"
-            | "LPT6"
-            | "LPT7"
-            | "LPT8"
-            | "LPT9"
-    )
-}
-
-/// Hard ceiling on the per-entry `Vec::with_capacity` hint. 64 KiB
-/// covers the bulk of real npm package files (JS / JSON / TS, all
-/// typically under a few KiB each) without trusting the declared
-/// header size, which an attacker controls. `read_to_end` grows
-/// past this ceiling when a legitimate larger file warrants it.
-const VEC_PREALLOC_CEILING: usize = 64 * 1024;
-
-/// A `Read` wrapper that refuses to deliver more than `remaining`
-/// bytes. Unlike `std::io::Read::take`, exhaustion produces an
-/// explicit `io::Error` rather than a clean EOF. When the wrapped
-/// reader is a gzip decoder feeding a tar archive, a clean EOF at
-/// a block boundary would let a crafted archive silently truncate
-/// into a partial index. Surfacing an error keeps the archive
-/// iterator from accepting a half-read stream as complete.
-pub(crate) struct CappedReader<R: std::io::Read> {
-    inner: R,
-    remaining: u64,
-}
-
-impl<R: std::io::Read> CappedReader<R> {
-    pub(crate) fn new(inner: R, cap: u64) -> Self {
-        Self {
-            inner,
-            remaining: cap,
-        }
-    }
-}
-
-impl<R: std::io::Read> std::io::Read for CappedReader<R> {
-    fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
-        // A zero-length read is a no-op by the `Read` contract and
-        // must not error even if the cap is already exhausted.
-        if buf.is_empty() {
-            return Ok(0);
-        }
-        if self.remaining == 0 {
-            return Err(std::io::Error::new(
-                std::io::ErrorKind::InvalidData,
-                format!(
-                    "tarball decompression exceeds archive cap of {MAX_TARBALL_DECOMPRESSED_BYTES} bytes"
-                ),
-            ));
-        }
-        let want = buf.len().min(self.remaining as usize);
-        let n = self.inner.read(&mut buf[..want])?;
-        self.remaining -= n as u64;
-        Ok(n)
-    }
-}
-
-/// Maximum total decompressed bytes accepted from a single tarball.
-/// 1 GiB. Reality check against the npm registry on 2026-04-19.
-/// Biggest tarball in the top 1000 by download count is `next` at
-/// 154 MiB unpacked. Second is `@tensorflow/tfjs` at 147 MiB. The
-/// cap sits ~6x above both, leaves room for future growth, and
-/// stays well below the process RSS a gzip bomb would otherwise
-/// force the installer to allocate.
-#[cfg(not(test))]
-pub(crate) const MAX_TARBALL_DECOMPRESSED_BYTES: u64 = 1 << 30;
-#[cfg(test)]
-pub(crate) const MAX_TARBALL_DECOMPRESSED_BYTES: u64 = 1 << 20;
-
-/// Maximum bytes for a single tar entry. 512 MiB. Reality check: the
-/// largest legitimate single file shipped by a top-1000 npm package
-/// sits in the tens of MiB range (bundled WASM blobs in `@swc/wasm`,
-/// `@babel/standalone`, `monaco-editor`). 512 MiB leaves a full
-/// order of magnitude of headroom.
-#[cfg(not(test))]
-pub(crate) const MAX_TARBALL_ENTRY_BYTES: u64 = 512 << 20;
-#[cfg(test)]
-pub(crate) const MAX_TARBALL_ENTRY_BYTES: u64 = 1 << 20;
-
-/// Maximum number of tar entries in a single archive. 200_000.
-/// Reality check: `next` ships 8_065 files and `@fluentui/react`
-/// ships 7_448, the largest counts in the top 1000. 200_000 is
-/// ~25x above that and stops a crafted archive from pinning the
-/// CPU on iteration alone.
-#[cfg(not(test))]
-pub(crate) const MAX_TARBALL_ENTRIES: usize = 200_000;
-#[cfg(test)]
-pub(crate) const MAX_TARBALL_ENTRIES: usize = 64;
-
-// Thin wrapper over posix_fallocate(3) which returns the error code
-// directly (does not set errno). Caller decides how to handle the
-// error. Existing call site uses `let _ = ...` to ignore EOPNOTSUPP /
-// ENOSYS / EINVAL on filesystems where pre-allocation is a no-op.
-//
-// `len` is `libc::off_t` because that's the type the underlying glibc
-// signature uses, and it varies per target: `i64` on 64-bit Linux and
-// on 32-bit Linux when the libc bindings opt into _FILE_OFFSET_BITS=64,
-// `i32` on 32-bit Linux otherwise (e.g. Debian/Ubuntu's armhf packaging
-// build env). Taking it directly keeps the call-site cast in one place.
-#[cfg(target_os = "linux")]
-fn posix_fallocate(file: &std::fs::File, len: libc::off_t) -> std::io::Result<()> {
-    use std::os::fd::AsRawFd;
-    if len <= 0 {
-        return Ok(());
-    }
-    // SAFETY: fd is owned by `file` for the duration of the call.
-    let r = unsafe { libc::posix_fallocate(file.as_raw_fd(), 0, len) };
-    if r == 0 {
-        Ok(())
-    } else {
-        Err(std::io::Error::from_raw_os_error(r))
-    }
-}
-
-// Unsupported means kernel/fs lacks O_TMPFILE, caller falls back.
-// Hard is a real I/O error that bubbles up.
-#[cfg(target_os = "linux")]
-enum OTmpfileFallback {
-    Unsupported,
-    Hard(Error),
-}
-
-// Size threshold below which we skip both `posix_fallocate` and
-// `posix_fadvise(DONTNEED)` on the CAS write path. Both are
-// fixed-cost-per-call best-effort advisory syscalls whose benefits
-// (avoid ext4 fragmentation, evict pages) don't apply to small
-// writes — the kernel won't fragment a single-block write, and tiny
-// pages don't meaningfully pressure the cache. samply on a cold
-// 1230-pkg install pinned the two at ~4.4% + ~4.8% of self time
-// before this gate; gating to ≥64KB skips them for >95% of npm
-// tarball entries while preserving the original behavior on the
-// large files (typescript.js, monaco-editor, etc.) where it pays.
-//
-// Overridable via `AUBE_CAS_SMALL_FILE_THRESHOLD` (bytes). Set to 0
-// to restore the always-on behavior; set to a very large number to
-// effectively disable both syscalls.
-#[cfg(target_os = "linux")]
-const CAS_SMALL_FILE_THRESHOLD_DEFAULT: usize = 64 * 1024;
-
-#[cfg(target_os = "linux")]
-fn cas_small_file_threshold() -> usize {
-    static THRESHOLD: std::sync::OnceLock<usize> = std::sync::OnceLock::new();
-    *THRESHOLD.get_or_init(|| match std::env::var("AUBE_CAS_SMALL_FILE_THRESHOLD") {
-        Err(_) => CAS_SMALL_FILE_THRESHOLD_DEFAULT,
-        Ok(raw) => raw.parse::<usize>().unwrap_or_else(|_| {
-            warn!(
-                "AUBE_CAS_SMALL_FILE_THRESHOLD={raw:?} is not a non-negative integer; \
-                 falling back to default {CAS_SMALL_FILE_THRESHOLD_DEFAULT}"
-            );
-            CAS_SMALL_FILE_THRESHOLD_DEFAULT
-        }),
-    })
-}
-
-// Open anonymous file in parent dir, write, linkat via /proc/self/fd.
-// Skips the tempfile unique-name probe and explicit fchmod. Falls
-// back via Unsupported on EOPNOTSUPP, ENOENT (no /proc), or EXDEV.
-// AUBE_DISABLE_O_TMPFILE forces the legacy path.
-#[cfg(target_os = "linux")]
-fn try_o_tmpfile_publish(path: &Path, bytes: &[u8]) -> Result<CasWriteOutcome, OTmpfileFallback> {
-    use std::ffi::CString;
-    use std::io::Write;
-    use std::os::fd::FromRawFd;
-    use std::os::unix::ffi::OsStrExt;
-
-    let parent = path.parent().ok_or(OTmpfileFallback::Hard(Error::Io(
-        path.to_path_buf(),
-        std::io::ErrorKind::NotFound.into(),
-    )))?;
-    let parent_c = CString::new(parent.as_os_str().as_bytes()).map_err(|_| {
-        OTmpfileFallback::Hard(Error::Io(
-            path.to_path_buf(),
-            std::io::Error::new(std::io::ErrorKind::InvalidInput, "parent path has nul"),
-        ))
-    })?;
-    // SAFETY: `parent_c` is valid for the duration of the call.
-    let raw_fd = unsafe {
-        libc::open(
-            parent_c.as_ptr(),
-            libc::O_TMPFILE | libc::O_RDWR | libc::O_CLOEXEC,
-            0o644 as libc::c_uint,
-        )
-    };
-    if raw_fd < 0 {
-        let err = std::io::Error::last_os_error();
-        return match err.raw_os_error() {
-            // Old kernels lack O_TMPFILE. Overlayfs/tmpfs return
-            // EOPNOTSUPP, EISDIR, or EINVAL on some kernels.
-            // ENOTSUP is the same value as EOPNOTSUPP on Linux.
-            Some(libc::EOPNOTSUPP) | Some(libc::EISDIR) | Some(libc::EINVAL) => {
-                Err(OTmpfileFallback::Unsupported)
-            }
-            _ => Err(OTmpfileFallback::Hard(Error::Io(path.to_path_buf(), err))),
-        };
-    }
-    // SAFETY: raw_fd is owned, OwnedFd closes on drop.
-    let owned = unsafe { std::os::fd::OwnedFd::from_raw_fd(raw_fd) };
-    let mut file = std::fs::File::from(owned);
-    let small_threshold = cas_small_file_threshold();
-    let is_large = bytes.len() >= small_threshold;
-    // Best-effort fallocate so the kernel allocates contiguous extents
-    // up front. Skips ext4 fragmentation churn on the next write.
-    // EOPNOTSUPP and ENOSYS are fine, regular write_all handles them.
-    // Skipped below `small_threshold`: fragmentation only matters for
-    // multi-block writes, and most npm tarball entries are well under
-    // that. See `cas_small_file_threshold` for rationale.
-    if is_large {
-        let _ = posix_fallocate(&file, bytes.len() as libc::off_t);
-    }
-    file.write_all(bytes)
-        .map_err(|e| OTmpfileFallback::Hard(Error::Io(path.to_path_buf(), e)))?;
-    // No sync_data: contradicts the no-fsync CAS policy. Crash window
-    // between write and linkat is acceptable, lockfile + state hash
-    // recovers the missing entry on next install.
-
-    let proc_link = format!("/proc/self/fd/{}", std::os::fd::AsRawFd::as_raw_fd(&file));
-    let proc_c = CString::new(proc_link.as_bytes()).map_err(|_| {
-        OTmpfileFallback::Hard(Error::Io(
-            path.to_path_buf(),
-            std::io::Error::other("fd path has nul"),
-        ))
-    })?;
-    let final_c = CString::new(path.as_os_str().as_bytes()).map_err(|_| {
-        OTmpfileFallback::Hard(Error::Io(
-            path.to_path_buf(),
-            std::io::Error::new(std::io::ErrorKind::InvalidInput, "path has nul"),
-        ))
-    })?;
-    // SAFETY: both CStrings live through the call. AT_SYMLINK_FOLLOW
-    // resolves the /proc/self/fd magic-link to the anon inode.
-    let r = unsafe {
-        libc::linkat(
-            libc::AT_FDCWD,
-            proc_c.as_ptr(),
-            libc::AT_FDCWD,
-            final_c.as_ptr(),
-            libc::AT_SYMLINK_FOLLOW,
-        )
-    };
-    if r == 0 {
-        // CAS bytes are read-once into reflinks/hardlinks. Drop them
-        // from the page cache so the parallel linker pass over many
-        // packages doesn't push the working set out. Per-file cost is
-        // roughly fixed regardless of size, so small files paid a
-        // disproportionate share — gate on `small_threshold` to match
-        // the fallocate gate above.
-        if is_large {
-            use std::os::fd::AsRawFd;
-            let fd = file.as_raw_fd();
-            // SAFETY: fd is still owned by `file` here. POSIX_FADV_DONTNEED
-            // is advisory, return value is ignored.
-            unsafe {
-                libc::posix_fadvise(fd, 0, 0, libc::POSIX_FADV_DONTNEED);
-            }
-        }
-        return Ok(CasWriteOutcome::Created);
-    }
-    let err = std::io::Error::last_os_error();
-    match err.raw_os_error() {
-        Some(libc::EEXIST) => Ok(CasWriteOutcome::AlreadyExisted),
-        // No /proc in this sandbox.
-        Some(libc::ENOENT) => Err(OTmpfileFallback::Unsupported),
-        // Kernel opens O_TMPFILE but rejects linkat from /proc/self/fd.
-        // ENOTSUP is same value as EOPNOTSUPP on Linux.
-        Some(libc::EOPNOTSUPP) | Some(libc::EXDEV) => Err(OTmpfileFallback::Unsupported),
-        // Seccomp-filtered containers (gVisor, strict k8s pod-security
-        // profiles) block linkat and return EPERM/EACCES. Fall through
-        // to the tempfile path instead of aborting the install.
-        Some(libc::EPERM) | Some(libc::EACCES) => Err(OTmpfileFallback::Unsupported),
-        _ => Err(OTmpfileFallback::Hard(Error::Io(path.to_path_buf(), err))),
-    }
-}
-
-/// Map a nibble (0–15) to its lowercase hex ASCII byte. Used by
-/// `ensure_shards_exist` to build the 256 two-character shard names
-/// without pulling in `format!`/`hex` per call.
-fn hex_digit(n: u8) -> u8 {
-    match n {
-        0..=9 => b'0' + n,
-        10..=15 => b'a' + n - 10,
-        _ => unreachable!(),
-    }
-}
-
-/// Validate a package name and return the `safe_name` form used as a
-/// cache filename stem (`/` collapsed to `__` so scoped names survive
-/// a single path component). Refuses anything outside the npm name
-/// grammar so a hostile packument cannot turn a cache write into an
-/// arbitrary-file-write primitive. Public so callers in
-/// `aube-registry` and `aube` (which own separate cache layouts under
-/// the same cache root) can share one validator.
-///
-/// A malicious packument can set `name` to `../../etc/passwd` (or, on
-/// Windows, to something with a drive prefix or backslash). The old
-/// `name.replace('/', "__")` only stripped forward slashes, so
-/// `index_dir().join(format!("{name}@{version}.json"))` would silently
-/// resolve outside the cache directory on the first resolve of the
-/// hostile package.
-///
-/// Accepted grammar is `[A-Za-z0-9_.-]` per component, with a single
-/// optional `@scope/` prefix. Uppercase and leading `.` / `_` are
-/// allowed on purpose: npm's registry bans them for *new* publishes
-/// but thousands of pre-rule packages (`JSONStream`, `Base64`, etc.)
-/// still resolve fine under pnpm and bun, and mirroring the registry's
-/// publish grammar here would block their cache path and break
-/// install. The only rejects are empty components, `.` / `..`, the
-/// 214-char length ceiling, and any byte outside the grammar.
-pub fn validate_and_encode_name(name: &str) -> Option<String> {
-    if name.is_empty() || name.len() > 214 {
-        return None;
-    }
-    let (scope, bare) = match name.strip_prefix('@') {
-        Some(rest) => {
-            let (s, b) = rest.split_once('/')?;
-            (Some(s), b)
-        }
-        None => (None, name),
-    };
-    let ok_component = |s: &str| -> bool {
-        // npm's registry bars new packages from leading `.` / `_` but
-        // historical packages that predate the rule still resolve
-        // fine, and scoped private registries allow them. Only bar
-        // empty and `.`/`..` since those collide with path components
-        // after the `/` → `__` folding.
-        if s.is_empty() || s == "." || s == ".." {
-            return false;
-        }
-        s.bytes()
-            .all(|b| b.is_ascii_alphanumeric() || matches!(b, b'-' | b'_' | b'.'))
-    };
-    if let Some(s) = scope
-        && !ok_component(s)
-    {
-        return None;
-    }
-    if !ok_component(bare) {
-        return None;
-    }
-    Some(name.replace('/', "__"))
-}
-
-/// Check a version string for use as a cache filename component.
-/// The lockfile already constrains versions to semver-ish shapes, but
-/// the cache path is independent of the lockfile on the write side so
-/// a crafted packument version would still land here. Returns `true`
-/// for anything the cache path builder is willing to accept.
-pub fn validate_version(version: &str) -> bool {
-    if version.is_empty() || version.len() > 256 {
-        return false;
-    }
-    // pnpm and bun sometimes route non-semver specs (git URLs, file
-    // specs, aliased registries) through the `version` slot, so the
-    // guard only needs to block what actually breaks the cache path
-    // builder: path separators on any platform, `\0`, control chars,
-    // and the two "this is a directory name" aliases.
-    if version
-        .bytes()
-        .any(|b| b.is_ascii_control() || matches!(b, b'/' | b'\\' | b'\0'))
-    {
-        return false;
-    }
-    if version == "." || version == ".." {
-        return false;
-    }
-    true
-}
-
-/// Verify that data matches an SRI integrity hash. Accepts any of
-/// `sha512-` / `sha384-` / `sha256-` / `sha1-` prefixed base64 digests
-/// — the set npm and pnpm accept in `dist.integrity`. Returns `Ok(())`
-/// on match, `Err(Error::Integrity)` on mismatch or unknown algorithm.
-pub fn verify_integrity(data: &[u8], expected: &str) -> Result<(), Error> {
-    let Some((algo, expected_b64)) = parse_sri(expected) else {
-        return Err(Error::Integrity(format!(
-            "unsupported integrity format (expected sha1/sha256/sha384/sha512-...): {expected}"
-        )));
-    };
-
-    // Stack-buffer the actual digest (max sha512 = 64 bytes) so the
-    // hot path stays allocation-free. sha512 reuses the thread-local
-    // hasher because it's the common case by 3+ orders of magnitude;
-    // the legacy algorithms one-shot a fresh hasher.
-    let mut actual_buf = [0u8; 64];
-    let actual_len = match algo {
-        IntegrityAlgo::Sha1 => {
-            let d = Sha1::digest(data);
-            actual_buf[..d.len()].copy_from_slice(&d);
-            d.len()
-        }
-        IntegrityAlgo::Sha256 => {
-            let d = Sha256::digest(data);
-            actual_buf[..d.len()].copy_from_slice(&d);
-            d.len()
-        }
-        IntegrityAlgo::Sha384 => {
-            let d = Sha384::digest(data);
-            actual_buf[..d.len()].copy_from_slice(&d);
-            d.len()
-        }
-        IntegrityAlgo::Sha512 => SHA512_HASHER.with(|cell| {
-            let mut hasher = cell.borrow_mut();
-            hasher.reset();
-            hasher.update(data);
-            let d = hasher.finalize_reset();
-            actual_buf[..d.len()].copy_from_slice(&d);
-            d.len()
-        }),
-    };
-    let actual = &actual_buf[..actual_len];
-
-    use base64::Engine;
-    let engine = base64::engine::general_purpose::STANDARD;
-    let mut expected_digest = [0u8; 64];
-    let matched = engine
-        .decode_slice(expected_b64, &mut expected_digest)
-        .map(|n| n == actual_len && expected_digest[..n] == actual[..])
-        .unwrap_or(false);
-    if matched {
-        Ok(())
-    } else {
-        let actual_b64 = engine.encode(actual);
-        Err(Error::Integrity(format!(
-            "integrity mismatch: expected {expected}, got {prefix}{actual_b64}",
-            prefix = algo.prefix(),
-        )))
-    }
-}
-
-/// Verify a precomputed SHA-512 digest against an SRI integrity
-/// string. Used by the streaming-tarball fetch path: SHA-512 is
-/// computed during the chunk read loop, then handed here so the
-/// owned `Bytes` are not re-hashed on the import side. Saves one
-/// pass over the buffer (~7 ms / 5 MB tarball).
-///
-/// Returns `Ok(true)` when the SRI uses SHA-512 and the digest
-/// matches. Returns `Ok(false)` when the SRI uses a non-SHA-512
-/// algo (legacy SHA-1 / SHA-256 / SHA-384) so the caller can
-/// fall through to the buffered `verify_integrity` path that
-/// re-hashes with the right algo. Returns `Err` on parse failure
-/// or SHA-512 mismatch.
-pub fn verify_precomputed_sha512(actual: &[u8; 64], expected: &str) -> Result<bool, Error> {
-    let Some((algo, expected_b64)) = parse_sri(expected) else {
-        return Err(Error::Integrity(format!(
-            "unsupported integrity format (expected sha1/sha256/sha384/sha512-...): {expected}"
-        )));
-    };
-    if !matches!(algo, IntegrityAlgo::Sha512) {
-        return Ok(false);
-    }
-    use base64::Engine;
-    let engine = base64::engine::general_purpose::STANDARD;
-    let mut expected_digest = [0u8; 64];
-    let decoded_len = match engine.decode_slice(expected_b64, &mut expected_digest) {
-        Ok(n) => n,
-        Err(e) => {
-            return Err(Error::Integrity(format!(
-                "integrity field has malformed base64: {expected} ({e})"
-            )));
-        }
-    };
-    if decoded_len != 64 {
-        return Err(Error::Integrity(format!(
-            "integrity field decoded to {decoded_len} bytes, expected 64 for sha512: {expected}"
-        )));
-    }
-    if expected_digest[..decoded_len] == actual[..] {
-        Ok(true)
-    } else {
-        let actual_b64 = engine.encode(actual);
-        Err(Error::Integrity(format!(
-            "integrity mismatch: expected {expected}, got sha512-{actual_b64}",
-        )))
-    }
-}
-
-/// Cross-check that an extracted tarball's `package.json` reports the
-/// same `name` and `version` the registry told us to fetch. This is the
-/// implementation behind the `strictStorePkgContentCheck` setting and
-/// guards against registry-substitution attacks where a tarball is
-/// served under one (name, version) but actually contains a different
-/// package on disk.
-///
-/// `index` must be the result of a freshly-completed `import_tarball`
-/// (or `import_directory`) — the helper reads `package.json` straight
-/// from the on-disk store path recorded in the index, so the bytes
-/// being validated are exactly the bytes that just landed in the CAS.
-///
-/// Returns `Ok(())` when both fields match, `Err(Error::PkgContentMismatch)`
-/// when they don't, and `Err(Error::Tar)` if the manifest is missing
-/// or unparseable. We deliberately treat a missing/broken manifest as
-/// a check failure rather than silently passing — a registry tarball
-/// without a usable `package.json` is itself a corruption signal.
-pub fn validate_pkg_content(
-    index: &PackageIndex,
-    expected_name: &str,
-    expected_version: &str,
-) -> Result<(), Error> {
-    // The two error paths below intentionally omit the
-    // `{expected_name}@{expected_version}` coordinate. Every caller
-    // wraps with `miette!("{name}@{version}: {e}")` (mirroring the
-    // Error::Integrity path), so embedding it here would print the
-    // same coordinate twice — same rationale as the
-    // Error::PkgContentMismatch return below.
-    let stored = index
-        .get("package.json")
-        .ok_or_else(|| Error::Tar("package.json missing from tarball".to_string()))?;
-    let bytes =
-        std::fs::read(&stored.store_path).map_err(|e| Error::Io(stored.store_path.clone(), e))?;
-    let v: serde_json::Value = sonic_rs::from_slice(&bytes)
-        .or_else(|_| serde_json::from_slice(&bytes))
-        .map_err(|e| Error::Tar(format!("invalid package.json: {e}")))?;
-    let actual_name = v.get("name").and_then(|n| n.as_str()).unwrap_or("");
-    let actual_version = v.get("version").and_then(|v| v.as_str()).unwrap_or("");
-    // Tolerate a leading `v` on the tarball's version (e.g. "v2.0.8").
-    // Some publishers ship this shape; npm and bun normalize it on
-    // install, so aube does too rather than rejecting a package the
-    // other managers accept. The registry-side coordinate is the
-    // source of truth, so we only normalize the tarball side.
-    let actual_version_normalized = actual_version
-        .strip_prefix('v')
-        .filter(|rest| rest.starts_with(|c: char| c.is_ascii_digit()))
-        .unwrap_or(actual_version);
-    // pnpm v9 lockfiles key git-hosted deps by the codeload tarball URL
-    // (or a `git+<url>#<commit>` form) in the `version` slot of the
-    // dep_path — that URL is what the resolver hands us as
-    // `expected_version`, and it can't meaningfully be compared to the
-    // tarball's real semver. pnpm scopes its equivalent check to
-    // registry sources; do the same by dropping the version comparison
-    // (but still checking the name) whenever `expected_version` isn't
-    // semver-shaped.
-    let expected_is_url_or_ref = expected_version.contains("://")
-        || expected_version.starts_with("git+")
-        || expected_version.starts_with("file:");
-    let version_matches = expected_is_url_or_ref || actual_version_normalized == expected_version;
-    if actual_name != expected_name || !version_matches {
-        // Only carry the *actual* coordinate the tarball declared.
-        // Every caller wraps the error with the expected
-        // `{name}@{version}: ` prefix (mirroring the Error::Integrity
-        // path), so embedding `expected` here would print the same
-        // coordinate twice in the rendered diagnostic.
-        return Err(Error::PkgContentMismatch {
-            actual: format!("{actual_name}@{actual_version}"),
-        });
-    }
-    Ok(())
-}
-
-/// Decode a pnpm-style SRI integrity string (`sha512-` / `sha384-` /
-/// `sha256-` / `sha1-` + base64) into its raw hex digest. Used by
-/// introspection commands that accept the registry integrity format
-/// as an ergonomic input, and by `index_path` to shard the cache
-/// directory by integrity prefix. Returns `None` if the input isn't a
-/// well-formed SRI integrity string.
-pub fn integrity_to_hex(integrity: &str) -> Option<String> {
-    let (_, b64) = parse_sri(integrity)?;
-    use base64::Engine;
-    let bytes = base64::engine::general_purpose::STANDARD.decode(b64).ok()?;
-    Some(hex::encode(bytes))
 }
 
 #[derive(Debug, thiserror::Error, miette::Diagnostic)]
